fix: error-boundary key invalidation
