+++ libs/pages/application/src/lib/ui/container/container.tsx
@@ -1,6 +1,6 @@
 import { type Environment } from 'qovery-typescript-axios'
 import { type PropsWithChildren, useContext } from 'react'
-import { useParams } from 'react-router-dom'
+import { useLocation, useParams } from 'react-router-dom'
 import { useCluster } from '@qovery/domains/clusters/feature'
 import { EnvironmentMode, useEnvironment } from '@qovery/domains/environments/feature'
 import { type AnyService, type Database } from '@qovery/domains/services/data-access'
@@ -8,14 +8,22 @@ import {
   NeedRedeployFlag,
   ServiceActionToolbar,
   ServiceAvatar,
+  ServiceStateChip,
   ServiceTemplateIndicator,
   ServiceTerminalContext,
   useService,
 } from '@qovery/domains/services/feature'
 import { VariablesProvider } from '@qovery/domains/variables/feature'
 import { IconEnum } from '@qovery/shared/enums'
-import { CLUSTER_URL } from '@qovery/shared/routes'
-import { Badge, ErrorBoundary, Header, Icon, Link, Section, Skeleton, Tooltip } from '@qovery/shared/ui'
+import {
+  APPLICATION_DEPLOYMENTS_URL,
+  APPLICATION_GENERAL_URL,
+  APPLICATION_SETTINGS_URL,
+  APPLICATION_URL,
+  APPLICATION_VARIABLES_URL,
+  CLUSTER_URL,
+} from '@qovery/shared/routes'
+import { Badge, ErrorBoundary, Header, Icon, Link, Section, Skeleton, type TabsItem, Tooltip } from '@qovery/shared/ui'
 import TabsFeature from '../../feature/tabs-feature/tabs-feature'
 
 export interface ContainerProps extends PropsWithChildren {
@@ -24,13 +32,50 @@ export interface ContainerProps extends PropsWithChildren {
 }
 
 export function Container({ children }: ContainerProps) {
-  const { organizationId = '', environmentId = '', applicationId = '' } = useParams()
+  const { organizationId = '', projectId = '', environmentId = '', applicationId = '' } = useParams()
   const { data: environment } = useEnvironment({ environmentId })
   const { data: service } = useService({ environmentId, serviceId: applicationId })
 
   const { setOpen } = useContext(ServiceTerminalContext)
 
   const { data: cluster } = useCluster({ organizationId, clusterId: environment?.cluster_id ?? '' })
+
+  const location = useLocation()
+
+  const tabItems: TabsItem[] = [
+    {
+      icon: <ServiceStateChip mode="running" environmentId={service?.environment?.id} serviceId={service?.id} />,
+      name: 'Overview',
+      active:
+        location.pathname ===
+        APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_GENERAL_URL,
+      link: APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_GENERAL_URL,
+    },
+    {
+      icon: <ServiceStateChip mode="deployment" environmentId={service?.environment?.id} serviceId={service?.id} />,
+      name: 'Deployments',
+      active:
+        location.pathname ===
+        APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_DEPLOYMENTS_URL,
+      link: APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_DEPLOYMENTS_URL,
+    },
+    {
+      icon: <Icon iconName="key" iconStyle="regular" />,
+      name: 'Variables',
+      active:
+        location.pathname ===
+        APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_VARIABLES_URL,
+      link: APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_VARIABLES_URL,
+    },
+    {
+      icon: <Icon iconName="gear" iconStyle="regular" />,
+      name: 'Settings',
+      active: location.pathname.includes(
+        APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_SETTINGS_URL
+      ),
+      link: APPLICATION_URL(organizationId, projectId, environmentId, applicationId) + APPLICATION_SETTINGS_URL,
+    },
+  ]
 
   const headerActions = (
     <div className="flex flex-row items-center gap-4">
@@ -91,10 +136,10 @@ export function Container({ children }: ContainerProps) {
               </ServiceTemplateIndicator>
             )}
           </Header>
-          <TabsFeature />
+          <TabsFeature items={tabItems} />
           <NeedRedeployFlag />
           <div className="mt-2 flex min-h-0 flex-grow flex-col items-stretch rounded-b-none rounded-t-sm bg-white">
-            <ErrorBoundary>{children}</ErrorBoundary>
+            <ErrorBoundary key={tabItems.find(({ active }) => active)?.link}>{children}</ErrorBoundary>
           </div>
         </Section>
       </ErrorBoundary>
