+++ libs/pages/services/src/lib/ui/container/container.tsx
@@ -255,7 +255,7 @@ export function Container({ children }: PropsWithChildren) {
           <Tabs items={tabsItems} contentRight={contentTabs} />
           {cancelOnGoing && <Banner color="yellow">Deployment cancel ongoing...</Banner>}
           <div className="mt-2 flex min-h-0 flex-grow flex-col items-stretch rounded-b-none rounded-t-sm bg-white">
-            <ErrorBoundary>{children}</ErrorBoundary>
+            <ErrorBoundary key={tabsItems.find(({ active }) => active)?.link}>{children}</ErrorBoundary>
           </div>
         </Section>
       </ErrorBoundary>
